Decouple (mostly) `class_field.rs` from `Attribute`
Summary:
In this diff we start to turn the corner in extracting class field related
logic from the `Attribute` code: we're able to modify all of the method
signatures in `class_field` to produce and work with `ClassAttribute` rather
than `Attribute`.

This is a minimal cutover diff; there are still a few messy things to clean up,
for example we can get rid of `Attribute.read_only_equivalent` shortly, and
the attribute subset code is messier than it will need to be once we make
the code aware that protocol attributes are always `ClassAttribute`.

I'll stack such cleanups on top.

Reviewed By: grievejia

Differential Revision: D81011740

fbshipit-source-id: 63a132de2c1a7e1301857b565470911e7720298e