Move dataclass converter hook
Summary:
This gets the dataclass converter logic - which is definitely class-field specific -
out of the general attribute setting code in `check_set_read_write_and_infer_narrow`
and into the `check_class_attr_set_and_infer_narrow` which is closer to where it
really belongs.

In addition to being more modular, this will save us a tiny bit of compute because
we pass slightly less data, skip a match, and do fewer allocations in the cases
that aren't class attributes.

Reviewed By: grievejia

Differential Revision: D81043781

fbshipit-source-id: 224a9fec5962bc60e93a7d8f64f24a147502511e