Move `ClassAttribute` to `class_field.rs`
Summary:
We did enough prep work separating the logic that this works pretty
well. The only really messy thing left is that the class attribute
set logic is currently using `AttributeBase1` to handle dataclass
converters.

Ideally, our class attribute representation will eventually be enough
to handle this in closed form, but I'll clean this up somewhat in the
next diff by just unwrapping the type information directly in `attr.rs`
and passing it down so that we don't have to make `AttributeBase1`
public.

Reviewed By: grievejia

Differential Revision: D81056125

fbshipit-source-id: d416f692ef09ee12a4a2c712305ef48cf75531da