Don't synthesize an internal error
Summary:
We can simplify the lookup in `is_protocol_subset_at_attr` by just
mapping, no need to synthesize a failed LookupResult just to throw
it away.

At this point it's also much clearer that our handling of `found` is
fishy... if we have a union and we found the attribute in some cases
but not others, we're dropping that information.

I'll add some tests and fix shortly.

Reviewed By: grievejia

Differential Revision: D81048538

fbshipit-source-id: b63080fef6c7b2a565130696778f1c108819c1ac