Inline `try_lookup_attr`
Summary:
One of the problems `attr.rs` developed as we hacked on it was a ton
of functions with subtly different names that did marginally different
things; `try_lookup_attr` was one of these (we've since removed some,
or otherwise consolidated).

At this point it only has one caller, the subset check for protocols.
If we just inline it into that function, we no longer have to deal with
a separate interface and a name that doesn't really capture what it
does and how that differs from other lookup functions.

We also can see that the logic is unnecessarily convoluted and there
is probably a bug handling unions, but inlining the code as-is is the
first step toward fixing those issues.

Reviewed By: grievejia

Differential Revision: D81048539

fbshipit-source-id: 20e91db5c1d243b435d654377aa93912a8ef8b90