Remove `resolve_instance_method`
Summary:
At one point this seemed to help, but given that there are 3 uses of
`resolve_as_instance_method` and two of them use instance attribute
lookup but the other uses quantified attribute lookup, any benefit
in reducing verbosity is outwieghed by the verbosity of the helper
itself and the fact that it introduces inconsistencies in abstraction
level.

All callers can instead just get a ClassAttribute and call
resolve_as_instance_method directly, this is clean and consistent and
requires less code overall.

Reviewed By: grievejia

Differential Revision: D81046508

fbshipit-source-id: 8a4de0076ef4262c650b8de8f435b2a4f256739b