Support extra args via class keywords
Summary:
I did something similar to mypy, which is to add a `*` param to the signature. This seems like a reasonable solution to me. There are a few changes in the unittest results:

- For aliases, we used to raise two errors when an alias/name is missing. One error was because the expected parameter is required, and the other is because the non-required keyword was unexpected.

Adding the extra `*` means we can expect anything, but since we already handled the requiredness correctly, we still catch the error.

This seems reasonable to me.

- For ConfigDict, we temporarily do not handle them correctly since we haven't handled the extra keyword there.

Also another note: I updated D81071525 with a comment to explain that the way we handle default arguments when both validation keys are true isn't fully matching runtime (mypy's approach also doesn't). To fully handle it, we need to generate overloads for init. This could be in v2 or we can revisit after we implement everything else, but it doesn't seem like the highest priority since mypy also doesn't fully match runtime for this.

Reviewed By: rchen152

Differential Revision: D81079015

fbshipit-source-id: a58be1b67757fa8c3602aaa3fd5fc4451b58915d