don't display inlay hint if we're calling a constructor
Summary:
partially addresses https://github.com/facebook/pyrefly/issues/836#issuecomment-3151193339

if we have something like `x = C()` and the inferred type for `x` is a class named `C`, I think we can assume that `C()` is a constructor.

this is a heuristic and won't work with qualified names, but maybe not a huge deal. We could try to do an ad-hoc solve on the function name & see if it's a `ClassDef` for the same type, but I'm not sure how expensive that would be. If you guys think that's better LMK

Reviewed By: stroxler

Differential Revision: D80974428

fbshipit-source-id: 02e1e7689332f9b04296f2444f16716f90592b76