Add an `is_debug` flag to Pyrefly's answers solver + type order
Summary:
A common problem in primarily print-debugging-oriented languages (like Rust) is
that if deep parts of the codebase don't have enough high-level information
to toggle fine-grained debug logging, it can become very hard to make behavior
observable because adding the necessary logging unconditionally may barf
huge amounts of irrelevant information.

My immediate motivation is that I believe is_subset_eq is failing to handle
`type[T]` for a quantified `T` in certain cases, and unconditionally dumping
information from the solver would lead to me seeing all the solves from stdlib,
which is unusable.

My current view is that fine-grained debugging is most cleanly done via a flag
that is global to the unit of debugging, which for answers solver is generally
the current thread (which allows us to let debug cross a module export boundary
when needed). This includes making the flag visible inside the solver, which is
what I need in my own case... it's easy for me to set the flag when I hit the
appropriate analysis in AnswersSolver but I need it to be ergonomic to then
use it for debug logging in `Solver`

Putting this up as an RFC because historically we were sometimes averse to
adding debug tools to prod code (especially on Pyre) but we've found this
sort of thing to be a huge unblock in Pyrefly (for example if we didn't add
the stack debug info, I think almost all of our cycle-breaking progress
would have been dramatically slower). This should be very close to zero-cost
in production since the methods are never called, it costs us exactly one ref
cell allocation per thread unless we actually use it.

Reviewed By: yangdanny97

Differential Revision: D80956824

fbshipit-source-id: 0e1809c29de899c36d8526266bfb32cd756a871a