Disable interpreter query and search path heuristics for buck
Summary:
When testing, I realized that Buck-check was using search path heuristics and querying for an interpreter when that information should be provided to us.

This diff disables that functionality.

# This diff stack...
works on emulating Buck's (and other build systems') file mappings into something that Pyrefly can model and work with.

We want to do this because:
- it will help Pyrefly work with build systems better in OSS
  - `pyrefly buck-check` exists, but assumes a specific model of how Pyrefly will be called and interact with build systems, which isn't necessarily how other build systems will work
- it will enable Pyrefly to work with build systems in the IDE

## The idea...
Model a virtual filesystem in Pyrefly. We can do this by taking from what Neil set up with in `pyrefly.bxl`, which is a way to query Buck for a mapping of `{<target>: {"srcs": {<module_name>: <module_path>}, "deps": <target>}`. This approach appropriately models the build system by keying everything on `Targets`, where *only* targets are unique, and files can appear as sources in multiple targets, and module names are also only unique *within a target*.

Given this information, we can have the sourcedb we provide to a `ConfigFile` able to understand this, and correctly understand that, given a module name, file, and target, which file the import is referring to.

With that information, it's only a matter of knowing when to re-query the sourcedb, and we should be able to replicate build systems in Pyrefly!

Reviewed By: grievejia

Differential Revision: D80854251

fbshipit-source-id: 4351e6f37beeb8cbf613a17a34b969f658365d50