Remove synthesized `__delitem__` method
Summary:
I realized while looking into adding `extra_items` handling to `__delitem__` that `__delitem__` for TypedDict is in a curious state:

* We special-case `del` operations, but we also have synthesized overloads to handle `__delitem__` being called directly. Therefore, we have essentially the same logic duplicated in two places.
* We don't have synthesized overloads for `__getitem__` or `__setitem__`, so our support level for calling magic methods directly is inconsistent.

We don't yet support argument type expansion for overloads, so switching over entirely to synthesizing overloads would cause problems. So the best way to de-dup our (increasingly complicated) get/set/del logic is to remove the overloads and rely entirely on the special-casing. It should be pretty rare to call a magic method directly.

Reviewed By: stroxler

Differential Revision: D80764667

fbshipit-source-id: a3c49a2b973423f861ba363bac50af9d7b81f23d