follow renamed export all the way
Summary:
fixes https://github.com/facebook/pyrefly/issues/971

This is what pylance does. Though it is confusing and might be nice to have an extra click in between, this is the most correct behavior.

Instead of stopping the search when we encounter an alised re-export, we should follow it until the final definition. This allows the hover text and types to work correctly. assignments do not have this behavior, only imports.

This requires keeping track of the old name in order to search for that name in the next module's exports which we do in `DefinitionStyle::ImportAs` and `ExportLocation::OtherModule`

Reviewed By: SamChou19815

Differential Revision: D81069092

fbshipit-source-id: 7dd34872f980490f0a1c59dc809a7601e1f16ccf