Move handle into pyrefly_python crate
Summary:
This diff only handles moving `handle.rs` into `pyrefly_build`. We want to move `Handle` into Pyrefly build, since it doesn't really need to be in `pyrefly/lib/state`, and I'd like to prevent us from leaking `Target` to anywhere else in Pyrefly.

With this, we can update a few places to query imports with the `Handle`, which will let us know which target we're querying an import from.

# This diff stack...
works on emulating Buck's (and other build systems') file mappings into something that Pyrefly can model and work with.

We want to do this because:
- it will help Pyrefly work with build systems better in OSS
  - `pyrefly buck-check` exists, but assumes a specific model of how Pyrefly will be called and interact with build systems, which isn't necessarily how other build systems will work
- it will enable Pyrefly to work with build systems in the IDE

## The idea...
Model a virtual filesystem in Pyrefly. We can do this by taking from what Neil set up with in `pyrefly.bxl`, which is a way to query Buck for a mapping of `{<target>: {"srcs": {<module_name>: <module_path>}, "deps": <target>}`. This approach appropriately models the build system by keying everything on `Targets`, where *only* targets are unique, and files can appear as sources in multiple targets, and module names are also only unique *within a target*.

Given this information, we can have the sourcedb we provide to a `ConfigFile` able to understand this, and correctly understand that, given a module name, file, and target, which file the import is referring to.

With that information, it's only a matter of knowing when to re-query the sourcedb, and we should be able to replicate build systems in Pyrefly!

Reviewed By: grievejia

Differential Revision: D80982988

fbshipit-source-id: 03a75f14bdbd3afd1291a79dea796c899f9e9ee4