Add a failing test case for union vs protocol
Summary:
We now handle unions much better in attribute lookup than we did: because
unions are valid bases, we correctly handle cases where a non-union unwraps
to a union (like a quantified bound to a union).

But there are still some edge cases making the erroneous layering assumptions
of the past - for example our union protocol logic is never correct unless the
union is at the top level, because we're trying to distribute the check over
unions and then ignore lookup failures which is not composeable, that only
works when all unions are at the top level.

We see this in the new test: we correctly reject a bad union directly assigned,
but accept a type var with the exact same union as a bound.

Reviewed By: grievejia

Differential Revision: D81051127

fbshipit-source-id: 810618f3abd7d6e87bbf8492880e0141318cb9e5