Fix bug when union shows up non-top-level for protocol
Summary:
To fix the handling of unions that aren't top-level, we just need to reject the
protocol check if `got` includes branches with attribute lookup failures.

----

Note: the conformance change is close to a no-op. We don't understand constrained
type variables correctly right now, which is the root cause - we treat `AnyStr`
as if it were bound by `bytes | str` and (correctly, given that wrong treatment)
complain that an operation like `+` isn't type-safe across the union.

This problem already existed, we just catch it in one more place and throw an
extra error, it's not a regression.

Reviewed By: grievejia

Differential Revision: D81051126

fbshipit-source-id: 6ff26d2fab47dda2fe08769b92288ab337e34c1f