Fix AttributeBase1::Quantified fallback behavior
Summary:
Before this diff, when the bound of a quantified type was not itself a class instance, we would fall back to treating the AttributeBase1::Quantified like the value type of the underlying type var. But this is wrong -- the value annotated by a type var is not itself a type var.

Fixes https://github.com/facebook/pyrefly/issues/921

Reviewed By: stroxler

Differential Revision: D81081035

fbshipit-source-id: 42b3b605a24ddd3be5821ed8687a3f6a53dd969b