Don't re-bind unbound callback protocols on lookup
Summary:
When a decorator returns a callable protocol, we unbind the protocol's bound method, returning a function without a self type.

Also, when we look up a function on a class, we then bind the function to the class.

Before this diff, we did both of these things together, meaning we tried to bind the receiver object to the parameter _after_ self from the callback protocol.

This diff checks the function kind before trying to bind. If we see a callback protocol, we know that the `self` parameter has already been removed, so we don't try to bind it again.

Fixes https://github.com/facebook/pyrefly/issues/809

Reviewed By: yangdanny97

Differential Revision: D81205240

fbshipit-source-id: 63c0c3ba3f3fb18cf164e03ae9bffab40f56ac1c