Distinguish between quantified types and values
Summary:
Type parameters are runtime objects and have runtime behavior. When used in static annotations, type checkers ascribe semantics to them as well.

Before this diff, we confused the runtime and static semantics. This is most evident in our handling of `P.args` and `P.kwargs` in attribute lookup.

Before this diff, the _value_ `P` in `P.args` had the type `Type::Type(Type::Quantified)`. Later, we improved support for accessing attributes on bound tparams, including their `type[T]` forms. This leads to a fundamental confusion.

This diff resolves the confusion by introducing a new `Type::QuantifiedValue` form. As a _value_ this type represents instances of the corresponding class type (TypeVar, ParamSpec, TypeVarTuple). As a _type_ it becomes just Type::Quantified.

This means that `type[T]` and `T`-as-a-value are no longer the same type representation. No more confusion.

Reviewed By: stroxler

Differential Revision: D81062729

fbshipit-source-id: addc22eada47ec38bcb46a7d781c1632db1aa18b