Extend init method generation to account for strictness
Summary:
The default for regular dataclasses vs pydantic models is different. For pydantic models, strictness is false, while for dataclasses, strictness is true.

Hence, it seemed natural to have a pyantic check in the answers phase. Although happy to hear any further thoughts on that.

The approach is to revert to the dynamic type as the annotation of the generated field when strictness is false (since that means we rely on runtime coersion)

Reviewed By: rchen152

Differential Revision: D81260976

fbshipit-source-id: 506697fdd22610cf3503371eb22e246c34b5cc37