Rename pydantic_model to pydantic_base_model
Summary: In pydantic, we have base models as well as root models (which inherit from base models). I think we should have a new boolean for root models and then rename the existing boolean to correctly mention that a given model is a base model.  The reason is that args to the root model should be positional rather than kwargs, so we need to handle it differently than base models that aren't root models.

Reviewed By: rchen152

Differential Revision: D81262856

fbshipit-source-id: c20c14aa1ced1e88498cb1fcaf8a7f095338babf