Add test covering self instantiation from literal receivers
Summary:
While working on Self typing, I ran into this question about literal types: Should we instantiate Self to the literal type, or should we use the class?

With a covariant-only Self type, we should pick the most precise type, `Literal[E.A]`. But since Self can appear contravariantly as well, choosing `Literal[E.A]` would overconstrain in the "binary operator" use case where contravariant `Self` commonly appears.

Indeed, both mypy and pyright choose the less precise type `E` as `Self`, so this test also documents behavior that is compatible with other checkers. This test documents that we should choose to promote literals.

Note that mypy distinguishes between annotated literal types and those arising from literal expressions, and will instantiate to the literal type in the former case. Currently, Pyrefly does not make this distinction, so the test documents behavior that is consistent with Pyrefly in general.

Reviewed By: stroxler

Differential Revision: D81273738

fbshipit-source-id: 68eb75e902b5dc9b139327e9855bf9de2194d2ab