Prefer module value over submodule
Summary:
It is always the case that if a name is defined explicitly
in `__init__.py`, it takes precedence over a submodule.

In prticular, the common pattern of `from .xyz import xyz`
needs to work or it causes major issues - for example, as
one of our tests demonstrated it was impossible to call
`unittest.main` in the idiomatic way without a Pyrefly error
prior to this fix.

Note that this does not mean we get every import-issue
scenario correct:
- Issue #1008 tracks that we fail to bind some names that are implicitly bound
  in `__init__.py` modules.
- We're still preferring submodules given an explicit import like
  `import package.xyz`, which is (perhaps surprisingly) not aligned
  with the runtime behavior... a new unit test is incoming for that.

This fixes #322.
It largely fixes #932

--------

The actual semantics on collisions that aren't of the form `from .xyz import xyz` can be ambiguous in ways we may never handle fully; I add a unit test to track this in D81257449 but I think we can consider the problem solved enough for now; I suspect real-world code doesn't hit the ambiguous case terribly often.

Reviewed By: grievejia

Differential Revision: D81257183

fbshipit-source-id: 488de645ff5cdc7f1dfcf0190b88c15500d3b2af