Add unit test of direct import when there's a shadow
Summary:
The semantics around name collisions are complicated, and in
general the final value of `pkg.xyz` depends on the order in
which imports are evaluated.

I'm not sure how feasible it is to actually model this; at the moment
we reasonably use the distinction between `import pkg.xyz`
and `from pkg import xyz` to approximate this, but that's only
an approximation.

One example where we get the wrong answer is `unittest.main`.

-------

I'm not sure what we should do long term here; one option would
be to:
(1) *always* prefer the attribute of `__init__.py` (which is needed
to get the `from .xyz import xyz` pattern, and that's too common
for us to get wrong)
(2) complain if `__init__` tries to define an attribute that shadows a subpackage, *unless* it is coming from a `from .xyz import xyz` statement

It's unclear how much this matters. The actual runtime behavior is not deterministic from a local-only view, so real codebases probably don't hit the ambiguous case very often.

Reviewed By: grievejia

Differential Revision: D81257449

fbshipit-source-id: b64ad0e0030bf17b545e2fce23a3af90a1aab5f0