narrow `getattr(...) is not None`
Summary:
This is kind of an edge case where people use something like `getattr(x, "attr") is not None` to check for the presence of an attribute.

It feels like a code smell though, so not sure if I should add this.

See examples of this in PyTorch: https://github.com/search?q=repo%3Apytorch%2Fpytorch+%22if+getattr%28%22&type=code&p=2

We don't narrow the type of the attribute here, only assert its presence.

Reviewed By: stroxler

Differential Revision: D81291847

fbshipit-source-id: f041aa236c294e201b5767f77a212f9aa55fd412