narrow `getattr`
Summary:
If the attribute does not exist and the default is falsy, we narrow to `Any`

If the attribute exists and the default is falsy, then we narrow it to truthy

If we're not sure the default is falsy, do nothing. If the default is truthy then we can't say that the attribute is present/truthy.

closes https://github.com/facebook/pyrefly/issues/684

Reviewed By: stroxler

Differential Revision: D81276291

fbshipit-source-id: 0ebb29125d7f12ff1b2cf4e6d6aedc6ce5660a50