Spawn transaction related threads using a single thread pool
Summary:
We have seen that the IDE server can spawn a lot of threads. These threads are supposed to be short-lived transactions, but they seem to get stuck sometimes, which leaves all of them open and use up a ton of resources.

This diff doesn't attempt to fix the underlying issue yet (possible deadlock?), but just to use a thread pool to limit the damage. In principle, it should be fine, because we cannot have multiple committable transactions running at the same time anyways.

Reviewed By: kinto0

Differential Revision: D81332344

fbshipit-source-id: e6f927467555b11c25cdbeb98fde45230b6869ed