Fix panic on `__all__` mutation w/o definition
Summary:
Pyrefly currently treats an augmented assignment as defining a name in the static
scope. This is a little strange since that's not how the runtime works, but it lest us
use `bind_name`, and we analyze it reasonably well since we still complain about the
name not being defined in the flow scope if there's no actual definition.

But it's also easy to mess up edge cases - here, we forgot to add
`__all__` to the definitions upon augmented assign, which led to code that mutated it
causing a Pyrefly crash when `bind_name` was called.

The fix is to add the name always, by using two ifs instead of
an if/else to handle `__all__` specific logic.

Fixes #1012

Reviewed By: grievejia

Differential Revision: D81330509

fbshipit-source-id: a83d9ddfe5d9b16135f300bfe4e5c670b743f64c