Update TSP to version 0.2.0 (#1010)
Summary:
This PR updates TSP to match the newest internal version that Pylance is using.

It eliminates the following requests:

- `SearchForTypeAttributes`
- `GetTypeAttributes`
- `GetSymbolsForFile`
- `GetSymbol`
- `GetFunctionParts`

It eliminates the `Attribute` class and replaces it with the already existing `Symbol` idea.

It adds two new requests

- `GetSymbolsForType` that can find the symbols for a type.
- `GetSymbolsForNode` that can find the symbols scoped by a node.

These two new requests handle the situations required by the 4 of the original functions. `GetFunctionParts` was eliminated entirely as it was printing out parameters and such and could be reimplemented with just the `GetRepr` request.

This PR is a prerequisite for adding the handlers for these requests.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1010

Reviewed By: stroxler

Differential Revision: D81353800

Pulled By: kinto0

fbshipit-source-id: 656fab5331dba768ede9cf830e3787bd300f9ef6