Change instantiation helper to modify &mut Type instead of transfering ownership
Summary:
In the next change, I want to store a ClassType instead of Type for descriptors. Storing a ClassType makes it safer/easier to instantiate the descriptor's `__get__` and `__set__` method on access.

We need to instantiate the ClassType in `instantiate_helper`, but that code only knows how to deal with owned `Type` values. But if we use `&mut Type` instead, I can use `visit_mut` over the ClassType instead.

I added _mut variants to places where the existing ownership-transferring version was used elsewhere, and left the ownership-transferring code alone. I briefly looked to see if the remaining callers could easily be changed, but it didn't seem worthwhile.

Reviewed By: ndmitchell

Differential Revision: D81392047

fbshipit-source-id: 5ccc252469b480d34493f430a34dfc9101087cb6