Fix metaclass method binding
Summary:
When we look up a method from a metaclass, we still need to bind the method to the class. Before this diff, we were binding the method to the metaclass itself.

Correctly binding to the class means we can correctly instantiate metaclass calls where the `cls` param is something like `type[T]`. A notable example of this is `EnumMeta.__iter__`.

While this does fix a long-standing issue with Enum iteration, there is still some problem with classes extending `(str, Enum)` or `StrEnum`.
Ref https://github.com/facebook/pyrefly/issues/1009

Fixes https://github.com/facebook/pyrefly/issues/71

Reviewed By: yangdanny97

Differential Revision: D81392199

fbshipit-source-id: 5e85ca50f923ae07f4de42220f2741b605e4667e