Adjust how class tparams are pushed down
Summary:
We allow class methods and static methods to refer to the bound class type parameters. When used in this way, we interpret the methods as being generic themselves. Mechanically, we push the class type parameters down into the method.

This diff changes the code for pushing the class type parameters down. The change is mostly to prepare for the _next_ diff in this stack, but has some nice effects as well:

1. We push class tparams to the _front_ of the method tparam list, if the method is generic. This matches what mypy does and is (IMO) less surprising.
2. For an overloaded function, before we pushed the class tparams into _every_ overload if _any_ overload used one. Now we only push class tparams into the specific overloads that depend on class tparams.
3. The code does less `clone`ing, but fusing this process with instantiation.

Important note: this diff goes out of its way to preserve existing behavior with self-substitution. The existing behavior looks quite awkward in this diff, but the subsequent diffs remove the awkwardness.

Reviewed By: yangdanny97

Differential Revision: D81392086

fbshipit-source-id: f2288c693908229376901ed158ec7193a054653a