Add ClassBase::to_self_type
Summary:
When we create a self type for a class (no explicit targs), we "instantiate" -- use the class tparams as its args. After we substitute the self type in, we will instantiate the tparams in the next step.

Since we need AnswerSolver to instantiate a class, we just store the instantiated class type in our ClassBase enum directly.

Reviewed By: yangdanny97

Differential Revision: D81392077

fbshipit-source-id: 3844430a2422d1411112123199ee96b0d2a3a1aa