Prepare protocol matching for self substitution
Summary:
Context:

I am working toward a change where we substitute Self on attribute lookup. This was originally how it worked, but changed in D71863313, partly because subst-on-lookup interfered with protocol matching.

The issue from before is that we were substituting `Self` in a protocol with the protocol type itself, meaning contravariant occurrences of `Self` were leading to failures.

```
from typing import Protocol, Self
class P(Protocol):
    def f(self, x: Self):
        pass
class C:
    def f(self, x: Self):
        pass
x: P = C()
```

The solution is _more_ self substitution in lookup, not less! When we see `C <: P`, we should substitute `Self/C` when looking up attributes from `P`.

This diff prepares for that by creating a new `Instance` type that contains the appropriate type to substitute for `Self`. However, since we still don't actually substitute self during attribute lookup, the changes are benign.

Reviewed By: yangdanny97

Differential Revision: D81392064

fbshipit-source-id: e738292bdb1b4e834f5eaf24bd450df04875ffb3