Distribute union when checking `await` expression
Summary:
rock_union_intersection

This would be a quick&dirty way to workaround a backtracking issue, so [this post](https://fb.workplace.com/groups/1355816225674448/permalink/1419798159276254/) can be addressed for now. There might be a way to improve on error reporting here. But getting the type right can definitely be a good first step.

Reviewed By: yangdanny97, samwgoldman

Differential Revision: D81285558

fbshipit-source-id: 8652079a03d204c08c6148c3d1e38bea11e68f94