More rootmodel testcases
Summary:
RootModels appear to be fairly restricted, so we don't need to deeply analyze the class heirchy.

My understanding from trying these examples on runtime is that:

A RootModel only takes one user defined TArg. There's already a generic TArg in the first position of a rootmodel. So I think the behavior should be:

- If we have exactly two targs in our RootModel, record the second one

- If we have one targ, record it

- If we have more than two, error (this is already checked from the definition)

Things to look into:

When do we want to extract this data? Options are:
 - at the metadata level with an extra field for RootModel containing the targs we will use
- or at the dataclass level.

Reviewed By: rchen152

Differential Revision: D81532828

fbshipit-source-id: 463d6cedf073417d042eccd4835cd2bacf29c935