Export all global variable names
Summary:
Export all global variable names, since those can be marked as sources or sinks in Pysa. We need to verify that users are modeling existing global variables.

We iterate on top level statements to find global variables.
We could potentially use the `Definitions` logic, but this handles logic with `__all__` that we don't really care about, so I decided to keep it simple.

Reviewed By: ndmitchell

Differential Revision: D81586865

fbshipit-source-id: 3549783217210b8c54c8e8a1cfc69aabf02b23aa