Bind to self on super() attribute lookup
Summary:
The added test currently passes, but for the wrong reason. We unsafely allow `C` to subtype `Self@C` (https://github.com/facebook/pyrefly/issues/706).

The modified test is kind of interesting. It passes in mypy but fails in pyright. The root cause is that mypy converts Self to a type parameter on the method, making `A.f` a call to a generic function with tparam `Self <: A`. I'm open to matching this behavior, but for the time being, I'm just modifying the expectation on the test.

Reviewed By: yangdanny97

Differential Revision: D81504351

fbshipit-source-id: bef6cbb46e2b742dd00a7066932c42efc6fe2ec1