Add inheritance rootmodel testcase
Summary:
My plan for this was that in class_metadata_of, we calculate the rootmodel type as follows:

-  For a `cls`, we check `get_base_types_for_class` and check if any of them is a RootModel and get the arg for that.

- If not, check if any of the basemodels have a RootModel in the metadata and propagate that upwards

Otherwise, check if the class itself is a RootModel

However, I am not sure if this is 100% correct, as I wonder if `class_metadata_of` is needed for `get_base_types_for_class` to work correctly.

Or more generally, Does pydantic Metadata actually need to be its own binding that gets solved separately, rather than being glued to class_metadata?

I will think about that next.

Reviewed By: rchen152

Differential Revision: D81605853

fbshipit-source-id: e5a8895397b2e5c5744958913fc5d2eb3170199c