Disable type errors by default in IDEs
Summary:
In this diff, we will switch to the model where type errors are disabled for a given file by default, unless the file is covered by a config that doesn't disable type errors (explicit configs enable type errors by default).

The behavior can be overridden on the IDE side with `displayTypeErrors: "force-on"` or `displayTypeErrors: "force-off`. In these cases, type errors will always be hidden/shown, regardless of configs.

On the rust binary side, I made the fallback config (which should correspond to the case where a file doesn't map to any config) have `disable_type_errors_in_ide: Some(true)`, so that we can have the desired behavior described above.

On the vscode extension side, I only signal it through the status bar. It's still a TODO to nug users to turn it on like pylance, but we should be careful not making it too noisy. See (https://github.com/microsoft/pylance-release/issues/7343)

Reviewed By: kinto0

Differential Revision: D81344137

fbshipit-source-id: 48811f3e35b49fda244cd4c737f37408a0864175