Show error enabled/disabled status per file in status bar
Summary:
Building on top of the previous diff, this diff instead switches the status bar to display the type error status of the current active file.

Since the knowledge of the information only exists server side, we have to use a custom protocol to talk to the server. On the client, whenever config or active file changed, we will send a simple request to the server asking for the type checking status for the active file. The server will return 5 possible variants. The title of the status bar will just say whether it's on or off, but the hover message will explain why it's in this state with more details.

Reviewed By: kinto0

Differential Revision: D81603099

fbshipit-source-id: b75fcb3d42fea2e4898839002b18d1b7629ec986