Handle scoping correctly for class fields in body (#1024)
Summary:
Fixes https://github.com/facebook/pyrefly/issues/264.

This diff addresses the scoping issues of class fields in nested scopes.

When a class field is defined in a class body, the scope of the field should only be available in the immediate class block, unless for annotations. This is described in the Python doc at https://docs.python.org/3/reference/executionmodel.html#resolution-of-names

To address this issue, we check the flow style when looking up a variable. If the variable is defined as a class field, we apply appropriate checks to avoid the incorrect scoping rule from being applied.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1024

Reviewed By: migeed-z

Differential Revision: D81681451

Pulled By: yangdanny97

fbshipit-source-id: d186c027097411d71dffa255e1d406ec66f1044b