Warn if deprecated function is referred (#1011)
Summary:
This diff implements the last missing feature for deprecation warnings (https://github.com/facebook/pyrefly/issues/376): to warn when the deprecated function is referred (e.g. passed as function reference).

The implementation checks when typechecking an expression, whether the type of an expression is a deprecated function/bound method. If that's the case, a deprecated warning is generated.

If we only add this check, we might run into the issue that a deprecated function call would generate 2 deprecation warnings, one when the function is typechecked, and another when the call is typechecked.

To avoid the duplication, we avoid emitting the deprecation warning when typechecking a function call, instead we emit the deprecation warning when generating the call target.

As a consequence, the error message of deprecation warning is changed from "Calling" a deprecated function to "Reference to" a deprecated function. Suggestions are welcome for how to improve clarify of this error message.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1011

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

Rollback Plan:

Reviewed By: rchen152

Differential Revision: D81681196

Pulled By: yangdanny97

fbshipit-source-id: 8d106589c7ce85cf8daa3c0f5992401edb448531