Add functionality to compute RootModel type after class metadata has been populated
Summary:
We cannot do this during class metadata itself unfortunately, because it creates a cycle. We specifically do not want to reason about subscripts for base classes at the metadata level.

To solve this, I will revert the chage I made which extends rootmodel with a type, and then compute that info only after the metadata has been calculated and then inject that before we compute the dataclass init.

Reviewed By: rchen152

Differential Revision: D81617903

fbshipit-source-id: 880ee223caf0f701de3190e4f0373529a4c06071