Do not require kwonly arguments for the "root" argument in root models and make them pos args.
Summary:
I tried passing the argument itself to the dataclass init method to make it as agnostic as possible, but it seemed to do so, only to unwrap everything right away to extract the name.

Strangely, pyrefly already catches if we want to incorrectly override `root=<incompatible_type>`  in a class definition. It does not validate the the init method, hence the diffs above. So we can guarantee that root model won't be overridden incorrectly and that we are assigning it the right type. Would love to know if the above is intentional or a bug.

For now, the current approach seems to work, but happy to know if there are any further comments.

Reviewed By: rchen152

Differential Revision: D81650065

fbshipit-source-id: 8637db17307a78158cac3a23f25060da64620525