Make uninitialized class attributes accessible from the class object
Summary:
The existing behavior is really quite reasonable, but existing type checkers treat an uninitialized attribute on the class body as being accessible from the class object, and we should to.

We can separate the notions of accessible and initialized here -- we currently don't warn that this attributes is possibly-uninitialized, but this change doesn't preclude that error.

Fixes https://github.com/facebook/pyrefly/issues/647

Reviewed By: rchen152

Differential Revision: D81603683

fbshipit-source-id: 16e6a0b248539ef08bc008f0a1aa56d6c0e46406