Simplify require
Summary:
The Require enum tells us how much data we should keep around for a given module in state. For example, in check, we want to keep Errors for all the checked files, but just Exports for everything else. In the IDE, we need to keep around more information.

Before this change, we held the "default" Require value on the transaction (and state), and we stored the effective Require value for a given module in ModuleData.

When the require value for a module changes (to include more data, say), we now just compare the previous value with the new one. This simplification is possible because state is now transactional.

Reviewed By: SamChou19815

Differential Revision: D81733295

fbshipit-source-id: 051930db566d0ce7038b4b860dd64deb6f756ffc