Simplify passing handles with require
Summary:
Before this diff, we were passing a slice containing tuples of Handle and Require, but Require is always the same for every Handle. Instead, we can pass a slice of Handles and single Require.

This works, and it simplifies a bunch of code. But temporarily makes State's "run" functions awkward, as they take both a Require for the passed-in Handles, and a default Require for modules reached as dependencies of those Handles.

The next diff removes the awkwardness and simplifies things further.

Reviewed By: maggiemoss

Differential Revision: D81717002

fbshipit-source-id: 2c6aa00f0b126e710cd3bcd81214e470fbe3021c