Store `infer_with_first_use` on `Answer`
Summary: Previously the answer phase were reading `infer_with_first_use` from the module config on every solve, which created lots of contention on the shared config data structure. This diff makes it such that the answer phase caches the value of the flag on itself, eliminating the need for repeated config read.

Reviewed By: samwgoldman

Differential Revision: D81814896

fbshipit-source-id: 93b0054f07327dbb6c0efe71889454cb57966980